강민수 23.09.18.0700

+++ Assets/Scripts/NightScripts/Health.cs
@@ -7,13 +7,13 @@ public class Health : MonoBehaviour
     [Header("References")]
 
     [Header("Attributes")]
-    [SerializeField] private int hitPoints = 2;
+    [SerializeField] private float hitPoints = 2f;
     [SerializeField] private int currencyWorth = 50;
 
     private bool isDestroyed = false;
 
 
-    public void TakeDamage(int dmg){
+    public void TakeDamage(float dmg){
         hitPoints -= dmg;
 
         if (hitPoints <= 0 && !isDestroyed){
